fsl_frontend/fsl_backend/requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
Command==0.1.0
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
etelemetry==0.3.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.38.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
looseversion==1.0.3
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
nibabel==5.0.0
nipype==1.8.5
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
prov==2.0.0
psycopg2-binary==2.9.5
pycparser==2.21
pydot==1.4.2
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
rdflib==6.2.0
requests==2.28.2
scipy==1.10.0
simplejson==3.18.1
six==1.16.0
traits==6.3.2
urllib3==1.26.14
Werkzeug==2.2.2
wxPython==4.2.0
